requirements.txt
torch>=2.0
torchvision
facenet-pytorch
numpy
opencv-python
pillow
pytorch-msssim
streamlit
pandas
requests
python-dotenv
openai
tqdm
matplotlib
seaborn
pyyaml
python-multipart
wget